Move mtd and ubifs stuff to a separated package
